RF: Move default validation target stim to module file
Working on reducing size (number of lines) needed for basic validation demo.
